#13: refactoring most of data structures
