+++ src/i_CoordTranslate_factory.hpp
@@ -18,49 +18,19 @@
 
 */
 
-#ifndef __COORDSYSTEM__HPP___
-#define __COORDSYSTEM__HPP___
+#ifndef __TP_COORD_I_COORDTRANSLATE_FACTORY_HPP___
+#define __TP_COORD_I_COORDTRANSLATE_FACTORY_HPP___
 
-#include "m_hwmoves.hpp"
+#include "i_CoordTranslate_factory.hpp"
+#include "dto_CoordTranslateConfig.hpp"
+#include "i_CoordTranslate.hpp"
 #include <tpcommon/position.hpp>
-
 #include <nlohmann/json.hpp>
 
 #include <array>
 
 namespace tp {
 namespace coord {
-
-    using tp::motor::Steps;
-
-    struct CoordTranslateConfig {
-        std::string motorConfiguration; //: "corexy",
-        struct {
-            double x; //: -1.0,
-            double y; //: 1.0,
-            double z; //: 1.0
-            double t; //: 1.0
-        } scale;
-        struct {
-            double a; //: 82.0512820513,
-            double b; //: 82.0512820513,
-            double c; //: 200
-            double d; //: 100
-        } stepsPerMm;
-    };
-
-    CoordTranslateConfig getDefaults_CoordTranslateConfig();
-    void to_json(nlohmann::json& j, const CoordTranslateConfig& p);
-    void from_json(const nlohmann::json& j, CoordTranslateConfig& p);
-    std::ostream& operator<<(std::ostream& os, CoordTranslateConfig const& value);
-    std::istream& operator>>(std::istream& is, CoordTranslateConfig& value);
-    bool operator==(const CoordTranslateConfig& l, const CoordTranslateConfig& r);
-
-    class i_CoordTranslate {
-    public:
-        virtual Steps translate(const Position& pos) = 0;
-        virtual Position translate(const Steps& steps) = 0;
-    };
 
     std::shared_ptr<i_CoordTranslate> CoordTranslate_corexy_factory(const std::array<double, 4> stepsPerMM_, const Position scaleAxis = Position(1, 1, 1, 1));
     std::shared_ptr<i_CoordTranslate> CoordTranslate_simple_factory(const std::array<double, 4> stepsPerMM_, const Position scaleAxis = Position(1, 1, 1, 1));
