+++ src/i_MotorMoves.cpp
@@ -1,0 +1,51 @@
+/*
+
+    Raspberry Pi G-CODE interpreter
+    Copyright (C) 2018  Tadeusz Puźniakowski puzniakowski.pl
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+*/
+
+#include "i_MotorMoves.hpp"
+
+#include "i_Stepper.hpp"
+#include "i_Spindle.hpp"
+#include "i_Buttons.hpp"
+
+#include <atomic>
+#include <condition_variable>
+#include <cstring>
+#include <iostream>
+#include <mutex>
+#include <queue>
+#include <thread>
+#include <vector>
+
+#include <chrono>
+
+using std::this_thread::sleep_for;
+using std::this_thread::sleep_until;
+
+using std::chrono::duration_cast;
+using std::chrono::high_resolution_clock;
+using std::chrono::microseconds;
+using std::chrono::milliseconds;
+using std::chrono::steady_clock;
+
+namespace tp {
+namespace motor {
+
+} // namespace motor
+} // namespace tp
