+++ src/m_motor_rpi.cpp
@@ -69,7 +69,7 @@ int map_peripheral(struct bcm2835_peripheral* p)
         MAP_SHARED,
         p->mem_fd, // File descriptor to physical memory virtual file '/dev/mem'
         p->addr_p // Address in physical map that we want this memory block to expose
-    );
+        );
 
     if (p->map == MAP_FAILED) {
         throw std::runtime_error("map_peripheral failed");
