+++ tests/unit_gcd_commands.cpp
@@ -1,4 +1,5 @@
-#include "coordsystem.hpp"
+#include "i_CoordTranslate.hpp"
+#include "i_CoordTranslate_factory.hpp"
 #include "gcd_commands.hpp"
 
 #include "tdd_helpers.hpp"
