+++ tests/unit_gcd_engine_imgsave.cpp
@@ -1,4 +1,4 @@
-#include "coordsystem.hpp"
+#include "i_CoordTranslate.hpp"
 #include "gcd_engine.hpp"
 
 #include "tdd_helpers.hpp"
