+++ tests/unit_machine.cpp
@@ -1,6 +1,9 @@
-#include "coordsystem.hpp"
-#include "m_hwmoves.hpp"
+#include "i_CoordTranslate.hpp"
+#include "i_CoordTranslate_factory.hpp"
+
+#include "i_MotorMoves_factory.hpp"
 #include "machine.hpp"
+
 #include <tpcommon/img_img8.hpp>
 
 #include "tdd_helpers.hpp"
